+++ libft/ft_strncmp.c
@@ -12,32 +12,31 @@
 
 #include "libft.h"
 
-int ft_strncmp(const char * str1, const char * str2, size_t n)
+int	ft_strncmp(const char * str1, const char * str2, size_t n)
 {
-    int     i;
-    int     result;
-
-    i = 0;
-    result = 0;
+	size_t	i;
+	int	result;
+	
+	i = 0;
+	result = 0;
 	if (str1 == NULL && str2 == NULL)
 		return (0);
 	else if (str1 == NULL)
 		return (0 - str2[0]);
 	else if (str2 == NULL)
 		return (str1[0]);
-    while ((str1[i] != 0 && str2[i] != 0) && i < n)
-    {
+	while ((str1[i] != 0 && str2[i] != 0) && i < n)
+	{
 		result = str1[i] - str2[i];
-        if (result != 0)
-            return (result);
+		if (result != 0)
+			return (result);
 		i++;
-    }
-
+	}
 	if (str1[i] == '\0' && str2[i] != '\0')
 		result = 0 - str2[i];
 	else if (str1[i] != '\0' && str2[i] == '\0')
 		result = str1[i];
 	else if (str1[i] == '\0' && str2[i] == '\0')
 		result = 0;
-    return (result);
+	return (result);
 }
